Convert stylelint_spec to stylelint_generator_spec
